+++ DipesLinkDeviceTransfer/Program.InitDatabase.cs
@@ -26,7 +26,7 @@ namespace DipesLinkDeviceTransfer
 
         private async Task InitDataAsync(JobModel selectedJob)
         {
-            //  Console.WriteLine("Toi bat dau load");
+              Console.WriteLine("Toi bat dau load");
 
             if (_SelectedJob == null) return;
             if (selectedJob.CompareType == CompareType.Database)
@@ -39,7 +39,15 @@ namespace DipesLinkDeviceTransfer
                 // Save to List
                 _ListPrintedCodeObtainFromFile = databaseTsk.Result;
                 _ListCheckedResultCode = checkedResultTsk.Result;
-
+                SharedFunctions.PrintDebugMessage("So luong database: " + _ListPrintedCodeObtainFromFile.Count.ToString());
+                //foreach(var item in _ListPrintedCodeObtainFromFile)
+                //{
+                //    foreach(var i in item)
+                //    {
+                //        Console.Write(i.ToString()+",");
+                //    }
+                //    await Console.Out.WriteLineAsync("");
+                //}
                 Task a = Task.Run(() =>
                  {
                      //Console.WriteLine("Task a work");
@@ -67,7 +75,6 @@ namespace DipesLinkDeviceTransfer
                     if (_SelectedJob.CompareType == CompareType.Database)
                     {
                         await InitCompareDataAsync(_ListPrintedCodeObtainFromFile, _ListCheckedResultCode);
-
 #if DEBUG
                         await Console.Out.WriteLineAsync($"POD filter: {_CodeListPODFormat.Count}");
                         //  await Console.Out.WriteLineAsync("Complete load !");
@@ -107,122 +114,14 @@ namespace DipesLinkDeviceTransfer
             var pathBackupPrintedResponse = SharedPaths.PathPrintedResponse + $"Job{JobIndex + 1}\\" + jobModel.PrintedResponePath;
 
             // Init Databse from file, add index column, status column, and string "Feild"
-            List<string[]> tmp = await Task.Run(() => { return InitDatabase(pathDatabase); });
+            List<string[]> tmp = await Task.Run(() => { return SharedFunctions.InitDatabaseWithStatus(pathDatabase); });
 
             // Update Printed status
             if (jobModel.PrintedResponePath != "" && File.Exists(jobModel.DatabasePath) && tmp.Count > 1)
             {
-                await Task.Run(() => { InitPrintedStatus(pathBackupPrintedResponse, tmp); });
+                await Task.Run(() => { SharedFunctions.InitPrintedStatus(pathBackupPrintedResponse, tmp); });
             }
             return tmp;
-        }
-
-
-
-        /// <summary>
-        /// Load database
-        /// </summary>
-        /// <param name="path"></param>
-        /// <returns></returns>
-        private List<string[]> InitDatabase(string? path)
-        {
-            List<string[]> result = new(); // List result
-            if (!File.Exists(path))
-            {
-                _InitDataErrorList.Add(NotifyType.DatabaseDoNotExist);
-                return result;
-            }
-            try
-            {
-                using var reader = new StreamReader(path, Encoding.UTF8, true);
-                var rexCsvSplitter = path.EndsWith(".csv") ? new Regex(@",(?=(?:[^""]*""[^""]*"")*(?![^""]*""))") : new Regex(@"[\t]");
-                int lineCounter = -1;
-                int columnCount = 0;
-
-                if (reader.ReadLine() is not null)
-                    while (!reader.EndOfStream)
-                    {
-                        var readLineRes = reader.ReadLine(); 
-                        if (readLineRes is not null)
-                        {
-                            string[] line = rexCsvSplitter.Split(readLineRes).Select(x => Csv.Unescape(x)).ToArray();
-                            lineCounter++;
-                            if (lineCounter == 0)
-                            {
-                                // Create additional database columns
-                                var tmp = new string[line.Length + 2];
-                                tmp[0] = "Index"; // thêm cột Index ở đầu
-                                tmp[^1] = "Status"; // Thêm cột status ở cuối (^1 đại diện phần tử cuối)
-                                for (int i = 1; i < tmp.Length - 1; i++)
-                                {
-                                    tmp[i] = line[i - 1] + $" - Field{i}"; // Thêm chữ field vào cạnh column
-                                }
-                                columnCount = tmp.Length;
-                                result.Add(tmp);
-                            }
-                            else
-                            {
-                                var tmp = new string[columnCount];
-                                tmp[0] = "" + lineCounter;
-                                tmp[columnCount - 1] = "Waiting"; // thêm trạng thái waiting cho trường status
-                                for (int i = 1; i < tmp.Length - 1; i++)
-                                {
-                                    if (i - 1 < line.Length)
-                                    {
-                                        tmp[i] = line[i - 1];
-                                    }
-                                    else
-                                    {
-                                        tmp[i] = "";
-                                    }
-                                }
-                                result.Add(tmp);
-                            }
-                        }
-                    }
-            }
-            catch (IOException) { _InitDataErrorList.Add(NotifyType.CannotAccessDatabase); }
-            catch (Exception) { _InitDataErrorList.Add(NotifyType.DatabaseUnknownError); }
-            return result;
-        }
-
-        /// <summary>
-        /// Cập nhật trạng thái Printed cho database vừa load vào
-        /// </summary>
-        /// <param name="pathBackupPrinted"></param>
-        /// <param name="dbList"></param>
-        private void InitPrintedStatus(string pathBackupPrinted, List<string[]> dbList)
-        {
-            if (!File.Exists(pathBackupPrinted))
-            {
-                _InitDataErrorList.Add(NotifyType.PrintedResponseDoNotExist);
-                return;
-            }
-            try
-            {
-                using StreamReader reader = new(pathBackupPrinted, Encoding.UTF8, true);
-                int i = -1;
-                var rexCsvSplitter = pathBackupPrinted.EndsWith(".csv") ? new Regex(@",(?=(?:[^""]*""[^""]*"")*(?![^""]*""))") : new Regex(@"[\t]");
-                while (!reader.EndOfStream)
-                {
-                    i++;
-                    if (i == 0) // Leave out the first row
-                    {
-                        reader.ReadLine();
-                    }
-                    else
-                    {
-                        string line = Csv.Unescape(rexCsvSplitter.Split(reader.ReadLine())[0]);
-                        if (int.TryParse(line, out int index))
-                        {
-                            dbList[index][^1] = "Printed"; // Get rows by index and last column ^1 Updated with the content Printed
-                        }
-                    }
-                    //  await Task.Delay(1);
-                }
-            }
-            catch (IOException) { _InitDataErrorList.Add(NotifyType.CannotAccessPrintedResponse); }
-            catch (Exception) { _InitDataErrorList.Add(NotifyType.PrintedStatusUnknownError); }
         }
 
         #endregion  LOAD DB AND PRINTED LIST
@@ -321,6 +220,7 @@ namespace DipesLinkDeviceTransfer
 
         private void InitCompareData(List<string[]> dbList, List<string[]> checkedResultList)
         {
+            _NumberOfDuplicate = 0;
             HashSet<string> _ValidCheckedResultCodeSet = new();
             string validCond = ComparisonResult.Valid.ToString();
             int columnCount = 5;
