+++ DipesLinkWPF/Languages/LanguageModel.cs
@@ -1,12 +1,6 @@
-﻿using DipesLink.ViewModels;
-using SharedProgram.Models;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.IO;
 using System.Windows;
+using ViewModelBase = DipesLink.ViewModels.ViewModelBase;
 
 
 namespace DipesLink.Languages
